feat: add HTTP executor type (#36)
- New HTTP executor type, for making HTTP requests when stage is
executed. The HTTP executor follows a "sync pattern", which means that a
single request-response is all it takes for the stage execution to
complete. For comparison: the Semaphore executor follows an "async
pattern", where the initial request (that creates the Semaphore
workflow) is not enough to complete the stage execution, and we need to
use the reference ID on the response to monitor the status of the
execution.
- HTTP executor type allows you to define execution outputs by including
an `outputs` field in the response.
- To make adding more types of executors easier, we move the logic for
executors to the `pkg/executors` package. That package exposes a
`Executor` interface that is consumed by the workers that handle
executions – PendingExecutionsWorker and ExecutionPoller.